implementa filtro de deletados no repository

+++ services/use-cases/produtos/produto.ts
@@ -1,5 +1,6 @@
 import { ProdutoService } from "services/api/produtos/produto-service";
 import { TProduto, TSyncStatus } from "services/api/produtos/types";
+import { TEntityFilter } from "services/repository/config/repository";
 import { ProdutoRepository } from "services/repository/produtos/produto-repository";
 import { removeAttributeByKey } from "utils/array";
 export class ProdutoUseCase {
@@ -14,13 +15,11 @@ export class ProdutoUseCase {
     this.produtoRepository = produtoRepository;
   }
 
-  async getAll() {
-    const produtos = await this.produtoService.getAll();
-    await this.produtoRepository.save(produtos.data);
-    return produtos.data;
+  async getByFilter(filter?: TEntityFilter<TProduto>) {
+    return await this.produtoRepository.getByFilter(filter);
   }
 
-  async changeStatus(produto: TProduto, status?: TSyncStatus) {
+  async changeStatus(produto: TProduto, status: TSyncStatus) {
     await this.produtoRepository.save([{...produto, status}]);
   }
 
@@ -30,6 +29,8 @@ export class ProdutoUseCase {
 
   async delete(produto: TProduto) {
     await this.changeStatus(produto, 'deleted');
+    const produtos = await this.produtoRepository.getAll();
+    return produtos;
   }
 
   async getSync(): Promise<TProduto[]> {
@@ -47,17 +48,19 @@ export class ProdutoUseCase {
   }
 
   async sendSync(): Promise<void> {
-    const updatedProdutos = await this.produtoRepository.getByKey('status', 'updated');
-    const deletedProdutos = await this.produtoRepository.getByKey('status', 'deleted');
+    const updatedProdutos = await this.getByFilter({ status: 'updated' });
+    const deletedProdutos = await this.getByFilter({ status: 'deleted' });
 
     const updatedToRepository = removeAttributeByKey<TProduto>(updatedProdutos, 'status');
     const updated = removeAttributeByKey<TProduto>(updatedToRepository, 'indexed_id');
     const deleted = deletedProdutos.map(produto => produto.sku);
 
-    await this.produtoService.sendSync({ updated, deleted });
+    if (updated.length > 0 || deleted.length > 0) {
+      await this.produtoService.sendSync({ updated, deleted });
 
-    await this.produtoRepository.save(updatedToRepository);
-    await this.produtoRepository.delete(deleted);
+      await this.produtoRepository.save(updatedToRepository);
+      await this.produtoRepository.delete(deleted);
+    }
   }
 
   async sync(): Promise<TProduto[]> {
